Text.Ogmarkup

Parse and Generate

ogmarkup :: (IsString a, Monoid a) => String -> GenConf a -> Either ParseError a

From a String, parse and generate an output according to a generation configuration. The inner definitions of the parser and the generator implies the output type has to be an instance of the IsString and Monoid classes.

Generation Configuration

data GenConf a

A data type to carry out the generation configuration. In particular, it works well to define a Typography and some marker such as HTML tags.

Constructors

GenConf

Fields

typography :: Typography a

The Typography to use for the generation

documentTemplate :: Template a

storyTemplate :: Template a

asideTemplate :: Maybe a -> Template a

paragraphTemplate :: Template a

tellerTemplate :: Template a

dialogueTemplate :: a -> Template a

thoughtTemplate :: a -> Template a

replyTemplate :: Template a

betweenDialogue :: a

emphTemplate :: Template a

strongEmphTemplate :: Template a

authorNormalize :: Maybe a -> a

printSpace :: Space -> a

type Template a = a -> a

Typography

frenchTypo :: IsString a => Typography a

A proposal for tho French typography. It can be used with several generation approach, as it stay very generic. Required the output type to be an instance of IsString.

englishTypo :: IsString a => Typography a

A proposal for tho English typography. It can be used with several generation approach, as it stay very generic. Required the output type to be an instance of IsString.

data Typography a

A Typography is a data type that tells the caller what space she should privileged before and after a text.

Constructors

Typography

Fields

decide :: Mark -> (Space, Space, a)

openDialogue :: Bool -> Maybe Mark

closeDialogue :: Bool -> Maybe Mark

Punctuation marks and space

data Mark

Mostly in order to deal with typographic spaces, main punctuation marks are tokenized during the parsing of an Ogmarkup document.

Constructors

Semicolon
The character ;

Colon
The character ,

Question
The character ?

Exclamation
The character !

OpenQuote
The character "

CloseQuote
The character "

Dash
The character – or the sequence --

LongDash
The character — or the sequence ---

Comma
The character ,

Point
The character .

Hyphen
The character -

SuspensionPoints
Two or more . or the character …

Apostrophe
The characters ' or ’

data Space

Deal with typographic spaces, especially when it comes to separate two texts. Because Space derives Ord, it is possible to use min and max to determine which one to use in case of a conflict.

Constructors

Normal
A normal space that can be turned into a newline for displaying.

Nbsp
A non breakable space, it cannot be turned into a newline.

None
No space at all.