Text.Ogmarkup.Private.Ast

Documentation

type Document a = [Section a]

A ogmarkup document internal representation waiting to be used in order to generate an output.

data Section a

Constructors

Story [Paragraph a]
The story as it goes

Aside (Maybe a) [Paragraph a]
Something else. Maybe a letter, a flashback, etc.

type Paragraph a = [Component a]

data Component a

Constructors

Teller [Format a]
A narrative description

Dialogue (Reply a) (Maybe a)
A dialogue reply

Thought (Reply a) (Maybe a)
Inner dialogue of the character.

data Reply a

A character line of dialogue. A reply may contain a descriptive part, which is not part of what the character actually says or thinks. We call the latter a "with say" reply untill someone gives use a better name for it.

Constructors

Simple [Format a]
A reply of the form: "Good morning."

WithSay [Format a] [Format a] [Format a]
A reply of the form: "Good morning," she says. "How are you?"

data Format a

A nested formatted text

Constructors

Raw [Atom a]
No particular emphasis is required on this sequence

Emph [Format a]
Surrounded by *.

StrongEmph [Format a]
Surrounded by **.

Quote [Format a]

data Atom a

An Atom is the atomic component of a Ogmarkup document. It can be either a punctuation mark or a word, that is a string.

Note that, by construction, OpenQuote and CloseQuote are not valid Mark values here. Indeed, they are implicit with the Quote constructor. This choice allows to enforce that an opened quote needs to be closed.

Constructors

Word a
A wrapped string

Punctuation Mark
A punctuation mark

data Mark

Mostly in order to deal with typographic spaces, main punctuation marks are tokenized during the parsing of an Ogmarkup document.

Constructors

Semicolon
The character ;

Colon
The character ,

Question
The character ?

Exclamation
The character !

OpenQuote
The character "

CloseQuote
The character "

Dash
The character – or the sequence --

LongDash
The character — or the sequence ---

Comma
The character ,

Point
The character .

Hyphen
The character -

SuspensionPoints
Two or more . or the character …

Apostrophe
The characters ' or ’