Text.Ogmarkup.Private.Config

Documentation

type Template a = a -> a

data GenConf a

A data type to carry out the generation configuration. In particular, it works well to define a Typography and some marker such as HTML tags.

Constructors

GenConf

Fields

typography :: Typography a

The Typography to use for the generation

documentTemplate :: Template a

storyTemplate :: Template a

asideTemplate :: Maybe a -> Template a

paragraphTemplate :: Template a

tellerTemplate :: Template a

dialogueTemplate :: a -> Template a

thoughtTemplate :: a -> Template a

replyTemplate :: Template a

betweenDialogue :: a

emphTemplate :: Template a

strongEmphTemplate :: Template a

authorNormalize :: Maybe a -> a

printSpace :: Space -> a