Text.Ogmarkup.Private.Generator

The Generator Monad

newtype Generator a x

The Generator Monad is eventually used to generate an output from a given 'Ast.Document. Internally, it keeps track of the previous processed Atom in order to deal with atom separation.

Constructors

Generator

Fields

getState :: StateT (a, Maybe (Atom a)) (Reader (GenConf a)) x

runGenerator

Arguments

:: Monoid a

=> Generator a x
The Generator to run

-> GenConf a
The configuration to use during the generation

-> a
The output

Run a Generator monad and get the generated output. The output type has to implement the class Monoid because the Generator monad uses the mempty constant as the initial state of the output and then uses mappend to expand the result as it process the generation.

Low-level Generators

askConf

Arguments

:: (GenConf a -> b)
The function to apply to the GenConf variable To retreive the wanted parameter.

-> Generator a b

Retreive a configuration parameter. Let the output untouched.

apply

Arguments

:: Monoid a

=> Template a
The Template to apply.

-> Generator a x
The Generator to run.

-> Generator a ()

Apply a template to the result of a given Generator before appending it to the previously generated output.

reset :: Generator a ()

Forget about the past and consider the next Atom is the first to be processed.

raw

Arguments

:: Monoid a

=> a
A sub-output to append

-> Generator a ()

Append a new sub-output to the generated output

AST Processing Generators

atom :: Monoid a => Atom a -> Generator a ()

Process an Atom and deal with the space to use to separate it from the paramter of the previous call (that is the previous processed Atom).

maybeAtom :: Monoid a => Maybe (Atom a) -> Generator a ()

Call atom if the parameter is not Nothing. Otherwise, do nothing.

atoms :: Monoid a => [Atom a] -> Generator a ()

Process a sequence of Atom.

format :: Monoid a => Format a -> Generator a ()

Process a Format.

formats :: Monoid a => [Format a] -> Generator a ()

Process a sequence of Format.

reply :: Monoid a => Maybe (Atom a) -> Maybe (Atom a) -> Reply a -> Generator a ()

Process a Reply.

component

Arguments

:: Monoid a

=> Bool
Was the last component an audible dialog?

-> Bool
Will the next component be an audible dialog?

-> Component a
The current to process.

-> Generator a ()

Process a Component.

paragraph :: Monoid a => Paragraph a -> Generator a ()

Process a Paragraph and deal with sequence of Reply.

paragraphs :: Monoid a => [Paragraph a] -> Generator a ()

Process a sequence of Paragraph.

section :: Monoid a => Section a -> Generator a ()

Process a Section.

sections :: Monoid a => [Section a] -> Generator a ()

Process a sequence of Section.

document :: Monoid a => Document a -> Generator a ()

Process a Document, that is a complete Ogmarkup document