Text.Ogmarkup.Private.Parser

Description

This module provides several parsers that can be used in order to extract the Ast of an Ogmarkup document.

Please consider that only document should be used outside this module.

Documentation

data ParserState

Keep track of the currently opened formats.

Constructors

ParserState

Fields

parseWithEmph :: Bool

Already parsing text with emphasis

parseWithStrongEmph :: Bool

_Already parsing text with strong emphasis

parseWithinQuote :: Bool

_Already parsing a quote

enterEmph :: OgmarkupParser ()

leaveEmph :: OgmarkupParser ()

enterStrongEmph :: OgmarkupParser ()

leaveStrongEmph :: OgmarkupParser ()

enterQuote :: OgmarkupParser ()

leaveQuote :: OgmarkupParser ()

initParserState :: ParserState

type OgmarkupParser = GenParser Char ParserState

An ogmarkup parser processes Char tokens and carries a ParserState

parse :: OgmarkupParser a -> String -> String -> Either ParseError a

withEmph :: OgmarkupParser Bool

withStrongEmph :: OgmarkupParser Bool

withinQuote :: OgmarkupParser Bool

document :: IsString a => OgmarkupParser (Document a)

See Document.

section :: IsString a => OgmarkupParser (Section a)

See Section.

aside :: IsString a => OgmarkupParser (Section a)

See Aside.

story :: IsString a => OgmarkupParser (Section a)

See Story.

paragraph :: IsString a => OgmarkupParser (Paragraph a)

See Paragraph.

component :: IsString a => OgmarkupParser (Component a)

See Component.

teller :: IsString a => OgmarkupParser (Component a)

See Teller.

dialogue :: IsString a => OgmarkupParser (Component a)

See Dialogue.

thought :: IsString a => OgmarkupParser (Component a)

See Thought.

talk

Arguments

:: IsString a

=> Char
A character to mark the begining of a reply

-> Char
A character to mark the end of a reply

-> (Reply a -> Maybe a -> Component a)
Either Dialogue or Thought according to the situation.

-> OgmarkupParser (Component a)

talk c c' constr wrap a reply surrounded by c and c' inside constr (either Dialogue or Thought).

authorName :: IsString a => OgmarkupParser a

reply :: IsString a => Char -> Char -> OgmarkupParser (Reply a)

reply parses a Reply.

format :: IsString a => OgmarkupParser (Format a)

See Format.

raw :: IsString a => OgmarkupParser (Format a)

See Raw.

emph :: IsString a => OgmarkupParser (Format a)

See Emph.

strongEmph :: IsString a => OgmarkupParser (Format a)

See StrongEmph.

quote :: IsString a => OgmarkupParser (Format a)

atom :: IsString a => OgmarkupParser (Atom a)

See Atom.

word :: IsString a => OgmarkupParser (Atom a)

See Word. This parser does not consume the following spaces, so the caller needs to take care of it.

longword :: IsString a => OgmarkupParser (Atom a)

Wrap a raw string surrounded by ` inside a Word.

>>> parse longword "" "`test *ei*`"
Right (Ast.Word "test *ei*")

Therefore, ` can be used to insert normally reserved symbol inside a generated document.

mark :: OgmarkupParser (Atom a)

See Punctuation. Be aware that mark does not parse the quotes because they are processed quote.

openQuote :: OgmarkupParser ()

See OpenQuote. This parser consumes the following blank (see blank) and skip the result.

closeQuote :: OgmarkupParser ()

See CloseQuote. This parser consumes the following blank (see blank) and skip the result.

asideSeparator :: OgmarkupParser ()

An aside section (see Aside) is a particular region surrounded by two lines of underscores (at least three). This parser consumes one such line.

blank :: OgmarkupParser ()

This parser consumes all the white spaces until it finds either an aside surrounding marker (see Aside), the end of the document or one blank line. The latter marks the end of the current paragraph.

skip :: OgmarkupParser a -> OgmarkupParser ()

skip p parses p and skip the result