Text.Ogmarkup.Private.Typography

Inner spaces representation

data Space

Deal with typographic spaces, especially when it comes to separate two texts. Because Space derives Ord, it is possible to use min and max to determine which one to use in case of a conflict.

Constructors

Normal
A normal space that can be turned into a newline for displaying.

Nbsp
A non breakable space, it cannot be turned into a newline.

None
No space at all.

Typography definition

data Typography a

A Typography is a data type that tells the caller what space she should privileged before and after a text.

Constructors

Typography

Fields

decide :: Mark -> (Space, Space, a)

openDialogue :: Bool -> Maybe Mark

closeDialogue :: Bool -> Maybe Mark

beforeAtom :: Typography a -> Atom a -> Space

From a Typography, it gives the space to privilege before the input Text.

afterAtom :: Typography a -> Atom a -> Space

From a Typography, it gives the space to privilege after the input Text.

normalizeAtom :: Typography a -> Atom a -> a

Normalize the input in order to add it to a generated Text.

Ready-to-use Typography

frenchTypo :: IsString a => Typography a

A proposal for tho French typography. It can be used with several generation approach, as it stay very generic. Required the output type to be an instance of IsString.

englishTypo :: IsString a => Typography a

A proposal for tho English typography. It can be used with several generation approach, as it stay very generic. Required the output type to be an instance of IsString.